improve group and property handling
Prior to this change,  property processing will throw if the association is not
found in the item list. That can happen because the assocation is broken, and can
also happen on a valid file if the property is associated with a group rather than
a single item.

This change links the property to either the item or the group, and protects
against the case where the association is not valid against either.

It also adds display of groups (including entity IDs, and properties) to the
filereader UI.