remove codes from Box-classes (NEEDS REVIEW)
in the javascript source (see commit 43c8b7ad600a99296902da1209abe5fc9d47092b) createBoxCtor adds the given type-argument to `BoxParser.boxCodes.push(type);`, similar pushes were commented out in other createXBoxCtor and still present in createSampleEntry.

this was ported to a call in the constructor, but this is not functionally identical: the calls in the factory-functions were only called when declaring the class, not during each instance creation.

currently I removed this functionality and did not replace it, if we want to re-introduce this functionality I would implement it via exports.