Fix forbidden bit check in av1C parsing
I'm not a 100% familiar with AV1 but I think this is how it is supposed to be fixed.

`(tmp >> 7) & 0x1 !== 1`

First of all this is wrong, `0x1 !== 1` has a priority over the `&` so this is always false and the error never occurs. I think here we are checking the forbidden bit which is the first bit of the byte and must always be 0. So we need `(byte >> 7 & 1) !== 0` to trigger the error (or perhaps `== 1`)