fix: parse meta box and related boxes in in QTFF format (#459)
* fix: Parse meta box differently in QTFF format

In the QTFF format (ie: Quicktime .MOV files) the 'meta' atom
is a container box instead of a full box. We probe 'meta' to
understand if it's a QTFF or ISO 'meta' box.

fixes #309

* feat: Parse other QTFF metadata atoms as boxes

This includes these atoms/boxes:
 * 'keys' which are the named keys in a QTFF 'meta' box
 * 'ilst' which typically follows the 'keys' atom within 'meta'
 * 'data' usually the the contents of an ilst atom

The 'data' box has many data types. Each type is described by the
enumeration dataBox.Types. Only support for those that have a test
file have been implemented so far:

 * UTF8
 * BE_UNSIGNED_INT
 * BE_SIGNED_INT
 * BE_FLOAT32

---------

Co-authored-by: DenizUgur <DenizUgur@users.noreply.github.com>